remove CFB as blockcipher streammode
the cfb mode can be used to create a streamcipher, but it requires
a keytype specified since the decryption, despite using the encrypt
transformation, works slightly different. this would break the current
streamcipher abstraction, which is designed to provide a single involutary
function for encryption and decryption.
once we add pure streamciphers which require the change of this
simplification, we can add cfb later again.

Signed-off-by: Janos Laube <janos.dev@gmail.com>